move getIOTABalanceInChain to waspclient
